Task-11: Refactored Kruskal's algorithm
