+++ task-11-minimum-spanning-tree/src/main/java/ru/otus/PrimsAlgorithm.java
@@ -1,5 +1,7 @@
 package ru.otus;
 
+import ru.otus.util.graph.Edge;
+
 import java.util.Arrays;
 
 public class PrimsAlgorithm {
