+++ task-11-minimum-spanning-tree/src/test/java/ru/otus/PrimsAlgorithmTest.java
@@ -1,6 +1,7 @@
 package ru.otus;
 
 import org.junit.jupiter.api.Test;
+import ru.otus.util.graph.Edge;
 
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
